0.21.4

+++ flutter_osm_interface/lib/src/osm_controller/osm_controller.dart
@@ -70,11 +70,11 @@ abstract class IBaseOSMController {
   );
 
   Future<void> setMinimumZoomLevel(
-    int minZoom,
+    double minZoom,
   );
 
   Future<void> setMaximumZoomLevel(
-    int maxZoom,
+    double maxZoom,
   );
 
   /// activate current location position
